Channel History and Debug Modes
Added debug mode for queues, user selection and power in beamforming.
This also have changes for channel history which can be used in the code
with SimStructs.chanHistory structure. Doppler feedback is also udpated
in the current code with SimParams.fbFraction which tells the fraction
of timing based on feedback.